Converted Filter UI to TypeScript
